pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "f5xc-api-enriched"
version = "0.1.0"
description = "F5 Distributed Cloud API specification enrichment and publishing system"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Robin Mordasiewicz"}
]
keywords = ["f5", "xc", "openapi", "api", "documentation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.32.0",
    "pyyaml>=6.0.2",
    "openapi-spec-validator>=0.7.2",
    "jsonschema>=4.23.0",
    "httpx>=0.28.0",
    "rich>=13.9.0",
    "deepdiff>=8.0.0",
    "language-tool-python>=2.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
f5xc-download = "scripts.download:main"
f5xc-enrich = "scripts.enrich:main"
f5xc-validate = "scripts.validate:main"
f5xc-merge = "scripts.merge_specs:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["scripts*", "config*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=scripts --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
